+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.
 - Command /rules
 - Event 232, 308 and 309 (numeric replies for /rules -- undocumented
   in the RFC)
+- Command /cycle
 
 ### Changed ###
 - On event join/part handle an empty user/host.